Planets (simple) data
A small module adapted from the overlay that allows to add the Alt/Az data on the 5 main planets in simplified (decimal degree) shot format and the info on which side East/West the planet is. This permit a very clean but effective visualization on the planet location (e.g. "Jupiter +42.3d E", "Saturn -8d W" etc)