Added RoI to temperature
Added a RoI field to limit the field of view taken into account for min/max/mean temperature reporting.
this does not impact the final image.
With the widefield MLX90640 you can pick up the temperature of the housing so this allows to remove that very high temp compared to the sky.

some minor formatting in an attempt to flake8