Update allsky_ina3221.py
modified the current calculation to remove the absolute value. This allows for easier tracking of current in and out of buses such as batteries.

added a power calculation in addition to the voltage and current.

added a timestamp to the json file being generated to track data freshness and assist in collecting the power data in other tools.